+++ src/Vortice.Direct3D11/Texture1DDescription.cs
@@ -28,7 +28,7 @@ namespace Vortice.Direct3D11
             int arraySize = 1,
             int mipLevels = 0, 
             BindFlags bindFlags = BindFlags.ShaderResource,
-            Usage usage = Usage.Default, 
+            ResourceUsage usage = ResourceUsage.Default, 
             CpuAccessFlags cpuAccessFlags = CpuAccessFlags.None, 
             ResourceOptionFlags optionFlags = ResourceOptionFlags.None)
         {